+++ packages/v1/shared/src/utils/json-schema.ts
@@ -258,6 +258,12 @@ export function convertJsonSchemaToZodSchema(
     let schema = z.object(spec).describe(jsonSchema.description);
     return required ? schema : schema.optional();
   } else if (jsonSchema.type === "string") {
+    if (jsonSchema.enum && jsonSchema.enum.length > 0) {
+      let schema = z
+        .enum(jsonSchema.enum as [string, ...string[]])
+        .describe(jsonSchema.description);
+      return required ? schema : schema.optional();
+    }
     let schema = z.string().describe(jsonSchema.description);
     return required ? schema : schema.optional();
   } else if (jsonSchema.type === "number") {