+++ packages/v1/a2ui-renderer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @copilotkit/a2ui-renderer
 
+## 1.52.2-next.1
+
+### Patch Changes
+
+- @copilotkit/react-core@1.52.2-next.1
+
+## 1.52.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [f1eaa4c]
+  - @copilotkit/react-core@1.52.2-next.0
+
 ## 1.52.1
 
 ### Patch Changes