+++ packages/v1/react-core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # ui
 
+## 1.52.2-next.1
+
+### Patch Changes
+
+- Updated dependencies [bf1fc6f]
+  - @copilotkit/shared@1.52.2-next.1
+  - @copilotkit/runtime-client-gql@1.52.2-next.1
+  - @copilotkitnext/core@1.52.2-next.1
+  - @copilotkitnext/react@1.52.2-next.1
+
+## 1.52.2-next.0
+
+### Patch Changes
+
+- f1eaa4c: Fixing a regression with setModalOpen
+  - @copilotkit/runtime-client-gql@1.52.2-next.0
+  - @copilotkit/shared@1.52.2-next.0
+  - @copilotkitnext/core@1.52.2-next.0
+  - @copilotkitnext/react@1.52.2-next.0
+
 ## 1.52.1
 
 ### Patch Changes