+++ packages/v2/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkitnext/core
 
+## 1.52.2-next.1
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.52.2-next.1
+
+## 1.52.2-next.0
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.52.2-next.0
+
 ## 1.52.1
 
 ### Patch Changes